backend/requirements.txt
Flask==3.0.0
plaid_python==28.0.0
python-dotenv==0.15.0
itsdangerous==2.1.2
werkzeug==3.1.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.3

# Testing dependencies
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-mock==3.12.0
